refactor(menu): remove stories directory
