+++ src/components/navigation/MenuLink/MenuLink.stories.tsx
@@ -1,6 +1,6 @@
 import { menuInvitationLinks } from "@pages/privileges/outlets/users/config/menuInvitation.config";
 import { BrowserRouter } from "react-router-dom";
-import { MenuLink, MenuLinkprops } from "..";
+import { MenuLink, MenuLinkprops } from ".";
 import { StoryFn } from "@storybook/react";
 
 const story = {
